last

+++ view.py
@@ -1,6 +1,6 @@
 def viewer(result):
     print(f"result = {result}")
 def get_int():
-    return int(input("value = "))
+    return float(input("value = "))
 def get_complex():
     return complex(int(input("value =")), int(input("value =")))
